+++ src/daemon/gateway.go
@@ -101,17 +101,43 @@ type Connection struct {
 	Height     uint64
 }
 
+// GetConnections returns all connections
+func (gw *Gateway) GetConnections() ([]Connection, error) {
+	var conns []Connection
+	var err error
+	gw.strand("GetConnections", func() {
+		conns, err = gw.getConnections(func(c gnet.Connection) bool {
+			return true
+		})
+	})
+	return conns, err
+}
+
 // GetOutgoingConnections returns solicited (outgoing) connections
 func (gw *Gateway) GetOutgoingConnections() ([]Connection, error) {
 	var conns []Connection
 	var err error
 	gw.strand("GetOutgoingConnections", func() {
-		conns, err = gw.getOutgoingConnections()
+		conns, err = gw.getConnections(func(c gnet.Connection) bool {
+			return c.Solicited
+		})
 	})
 	return conns, err
 }
 
-func (gw *Gateway) getOutgoingConnections() ([]Connection, error) {
+// GetIncomingConnections returns unsolicited (incoming) connections
+func (gw *Gateway) GetIncomingConnections() ([]Connection, error) {
+	var conns []Connection
+	var err error
+	gw.strand("GetIncomingConnections", func() {
+		conns, err = gw.getConnections(func(c gnet.Connection) bool {
+			return !c.Solicited
+		})
+	})
+	return conns, err
+}
+
+func (gw *Gateway) getConnections(flt func(c gnet.Connection) bool) ([]Connection, error) {
 	if gw.d.pool.Pool == nil {
 		return nil, nil
 	}
@@ -125,7 +151,7 @@ func (gw *Gateway) getOutgoingConnections() ([]Connection, error) {
 	conns := make([]Connection, 0, len(cs))
 
 	for _, c := range cs {
-		if c.Solicited {
+		if flt(c) {
 			conn := gw.newConnection(&c)
 			if conn != nil {
 				conns = append(conns, *conn)
@@ -849,9 +875,10 @@ func (gw *Gateway) GetAddressCount() (uint64, error) {
 
 // Health is returned by the /health endpoint
 type Health struct {
-	BlockchainMetadata visor.BlockchainMetadata
-	OpenConnections    int
-	Uptime             time.Duration
+	BlockchainMetadata  visor.BlockchainMetadata
+	OutgoingConnections int
+	IncomingConnections int
+	Uptime              time.Duration
 }
 
 // GetHealth returns statistics about the running node
@@ -865,15 +892,28 @@ func (gw *Gateway) GetHealth() (*Health, error) {
 			return
 		}
 
-		conns, err := gw.getOutgoingConnections()
+		conns, err := gw.getConnections(func(c gnet.Connection) bool {
+			return true
+		})
 		if err != nil {
 			return
 		}
 
+		outgoingConns := 0
+		incomingConns := 0
+		for _, c := range conns {
+			if c.Outgoing {
+				outgoingConns++
+			} else {
+				incomingConns++
+			}
+		}
+
 		health = &Health{
-			BlockchainMetadata: *metadata,
-			OpenConnections:    len(conns),
-			Uptime:             time.Since(gw.v.StartedAt),
+			BlockchainMetadata:  *metadata,
+			OutgoingConnections: outgoingConns,
+			IncomingConnections: incomingConns,
+			Uptime:              time.Since(gw.v.StartedAt),
 		}
 	})